src/processor/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "processor-engine"
version = "0.1.0"
description = "A timetable processing engine"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Nirman Doshi", email = "nirman123.doshi@gmail.com"}
]
keywords = ["timetable", "processor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "SQLAlchemy~=2.0.0",
    "img2table~=1.4.0",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.8.0",
    "Pillow>=12.0.0",
    "opencv-python>=4.12.0",
    "pdf2image>=1.17.0",
    "python-docx>=1.2.0",
    "docx2pdf>=0.1.8",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
]
